requirements.txt
# Smart Traffic Signal Control System
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pycocotools>=2.0.7
Pillow>=9.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
opencv-python>=4.7.0
tqdm>=4.65.0
tensorboard>=2.13.0
matplotlib>=3.7.0
gymnasium>=0.28.0
filterpy>=1.4.5
lapx>=0.5.5
stable_baselines3>=2.0.0
